Split test for SSL authentication with Multiplexed transports
