+++ tests/IceRpc.Quic.Tests/Transports/QuicTransportSslAuthenticationConformanceTests.cs
@@ -1,14 +1,18 @@
 // Copyright (c) ZeroC, Inc.
 
-using IceRpc.Conformance.Tests;
 using IceRpc.Tests.Common;
+using IceRpc.Transports;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 using System.Net.Quic;
+using System.Net.Security;
+using System.Security.Authentication;
+using System.Security.Cryptography.X509Certificates;
 
 namespace IceRpc.Tests.Transports;
 
-public class QuicTransportSslAuthenticationConformanceTests : MultiplexedTransportSslAuthenticationConformanceTests
+/// <summary>Test Ssl authentication with Quic transport.</summary>
+public class QuicTransportSslAuthenticationConformanceTests
 {
     [OneTimeSetUp]
     public void FixtureSetUp()
@@ -19,7 +23,93 @@ public class QuicTransportSslAuthenticationConformanceTests : MultiplexedTranspo
         }
     }
 
-    protected override IServiceCollection CreateServiceCollection() =>
+    [Test]
+    public async Task Ssl_client_connection_connect_fails_when_server_provides_untrusted_certificate()
+    {
+        // Arrange
+        await using ServiceProvider provider = CreateServiceCollection()
+            .AddSingleton(
+                new SslServerAuthenticationOptions
+                {
+                    ServerCertificate = new X509Certificate2("server-untrusted.p12"),
+                })
+            .AddSingleton(
+                new SslClientAuthenticationOptions
+                {
+                    RemoteCertificateValidationCallback = (sender, certificate, chain, errors) => false
+                })
+            .BuildServiceProvider(validateScopes: true);
+
+        var sut = provider.GetRequiredService<ClientServerMultiplexedConnection>();
+
+        var listener = provider.GetRequiredService<IListener<IMultiplexedConnection>>();
+
+        // Start the TLS handshake.
+        Task clientConnectTask = sut.Client.ConnectAsync(default);
+
+        // Act/Assert
+        Assert.That(
+            async () => await listener.AcceptAsync(default),
+            Throws.TypeOf<AuthenticationException>());
+        Assert.That(async () => await clientConnectTask, Throws.TypeOf<AuthenticationException>());
+    }
+
+    [System.Diagnostics.CodeAnalysis.SuppressMessage(
+        "Security",
+        "CA5359:Do Not Disable Certificate Validation",
+        Justification = "The client doesn't need to validate the server certificate for this test")]
+    [Test]
+    public async Task Ssl_server_connection_connect_fails_when_client_provides_untrusted_certificate()
+    {
+        // Arrange
+        await using ServiceProvider provider = CreateServiceCollection()
+            .AddSingleton(
+                new SslServerAuthenticationOptions
+                {
+                    ClientCertificateRequired = true,
+                    RemoteCertificateValidationCallback = (sender, certificate, chain, errors) => false,
+                    ServerCertificate = new X509Certificate2("server.p12"),
+                })
+            .AddSingleton(
+                new SslClientAuthenticationOptions
+                {
+                    ClientCertificates = new X509CertificateCollection()
+                    {
+                        new X509Certificate2("client-untrusted.p12")
+                    },
+                    RemoteCertificateValidationCallback = (sender, certificate, chain, errors) => true
+                })
+            .BuildServiceProvider(validateScopes: true);
+
+        var sut = provider.GetRequiredService<ClientServerMultiplexedConnection>();
+        var listener = provider.GetRequiredService<IListener<IMultiplexedConnection>>();
+
+        // Start the TLS handshake by calling connect on the client and server connections and wait for the
+        // connection establishment.
+        var clientConnectTask = sut.Client.ConnectAsync(default);
+
+        // Act/Assert
+        Assert.That(
+            async () =>
+            {
+                // The Quic listener internally kill the client connection if it's not valid.
+                using var cts = new CancellationTokenSource(TimeSpan.FromMilliseconds(250));
+                _ = await listener.AcceptAsync(cts.Token);
+            },
+            Throws.InstanceOf<OperationCanceledException>());
+
+        Assert.That(
+            async () =>
+            {
+                // The Quic client connection fails with AuthenticationException when try to create a stream.
+                await clientConnectTask;
+                var stream = await sut.Client.CreateStreamAsync(bidirectional: false, CancellationToken.None);
+                await stream.Output.WriteAsync(new ReadOnlyMemory<byte>(new byte[] { 0xFF }), CancellationToken.None);
+            },
+            Throws.TypeOf<AuthenticationException>());
+    }
+
+    private static IServiceCollection CreateServiceCollection() =>
         new ServiceCollection()
             .AddMultiplexedTransportTest(new Uri("icerpc://127.0.0.1:0/"))
             .AddQuicTransport();
